+++ CHANGELOG.md
@@ -6,12 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `transparent` bgcolor
 - `set_background` function
 - `is_valid_color` function
 - `color_complement` function
 - `select_color` function
 ### Changed
+- Transparent mode support for `bgcolor` parameter
 - Random mode modified
 - Complementary color support for `color` and `bgcolor` parameters
 - `filter_color` function modified